+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - Added CMCollections Resource
 - Added Set-ConfigMgrCert to the ResourceHelper
 - Added CMBoundaries resource
+- Added Convert-CidrToIP to the ResourceHelper
 
 ### Changed